UPDATE v1.2
Update Database: Add tables: QUOTE_TYPE | QUOTE | QUOTE_ITEM
Update API: TESTIMONIAL | SPONSOR | QUOTE

Note: When performing an update, if the QuoteItem in dto is less than in the current db, the remaining items in the db will be deleted. Otherwise, new items will be added. If the quantity is the same, only the value according to the item position in dto will be updated.